alterações feitas no controller e service

+++ src/main/java/com/lucaspires/crud_funcionarios/service/FuncionarioService.java
@@ -1,8 +1,11 @@
 package com.lucaspires.crud_funcionarios.service;
 
 import com.lucaspires.crud_funcionarios.entity.Funcionario;
+import com.lucaspires.crud_funcionarios.exceptions.RecursoNaoEncontrado;
 import com.lucaspires.crud_funcionarios.repository.FuncionarioRepository;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 
 @Service
@@ -15,22 +18,29 @@ public class FuncionarioService {
         this.repo = repo;
     }
 
-    public void salvarFuncionario(Funcionario funcionario){
-        repo.saveAndFlush(funcionario);
+    public Funcionario salvarFuncionario(Funcionario funcionario){
+       return repo.save(funcionario);
+    }
+
+    public List <Funcionario> listarFuncionarios(){
+        return repo.findAll();
     }
 
     public Funcionario buscarPorId(Integer id){
         return repo.findById(id).orElseThrow(
-                () -> new RuntimeException("Funcionario não econtrado"));
+                () -> new RecursoNaoEncontrado("Funcionario com ID "+id+" não econtrado"));
     }
 
     public void deletarPorId(Integer id){
+        if(!repo.existsById(id)){
+            throw new RecursoNaoEncontrado("Funcionário com ID "+id+" não encontrado.");
+        }
         repo.deleteById(id);
     }
 
     public void atualizarFuncionarioPorId(Integer id, Funcionario funcionario){
         Funcionario funcionarioEntity = repo.findById(id).orElseThrow(
-                () -> new RuntimeException("Funcionario não econtrado"));
+                () -> new RecursoNaoEncontrado("Funcionario com ID "+id+" não econtrado"));
         Funcionario funcionarioAtualizado = Funcionario.builder()
                 .nome(funcionario.getNome() != null ? funcionario.getNome() : funcionarioEntity.getNome())
                 .cargo(funcionario.getCargo() != null ? funcionario.getCargo() : funcionarioEntity.getCargo())
